refactor: Remove GradientText component(replace component with mixin)
