+++ src/components/ToggleButton/ToggleButton.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/_index.scss';
+@import '../../styles/index';
 
 .layout {
   display: flex;
