+++ src/utils/classNames.js
@@ -1,0 +1,3 @@
+export const cn = (...classNameList) => {
+  return classNameList.filter(className => !!className).join(' ');
+};
